+++ src/main/java/adapters/interfaces/IKeylistenner.java
@@ -2,7 +2,7 @@ package adapters.interfaces;
 
 import java.util.function.Consumer;
 
-import entities.key_event.KeyEvent;
+import entities.keyEvent.KeyEvent;
 
 public interface IKeylistenner {
     IKeylistenner setOnKeyPressedMethod(Consumer<KeyEvent> onKeyPressed);
