+++ src/main/java/usecases/actions/IRobot.java
@@ -1,4 +1,4 @@
-package usecases.actions_executor;
+package usecases.actions;
 
 import java.awt.datatransfer.Transferable;
 import java.awt.datatransfer.StringSelection;
